requirements.txt
build==0.10.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
packaging==23.0
platformdirs==3.1.1
pycryptodomex==3.17
pyproject-hooks==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
steam==1.4.4
tomli==2.0.1
urllib3==1.26.14
vdf==3.4
